Update ev3.py
Old code did not work because for a type mismatch and wrong values.
time.clock() referenced to a not well defined "processor time" which increases by approx 0.05 per second. Therefore time.time() should be used. A call to time.time() returns the system time as a float value in seconds. Documentation states that it may be very unprecise on some systems but it seems to work on the EV3.